refactor: add common page, update all tx type and util imports
