+++ packages/common/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+  "extends": ["plugin:require-extensions/recommended", "../../.eslintrc.json"],
+  "plugins": ["require-extensions"]
+}
