+++ packages/common/.npmignore
@@ -1,0 +1,11 @@
+*.log
+*.env
+tests
+src
+exports
+node_modules
+.eslintrc.json
+.turbo
+package-lock.json
+tsconfig.json
+index.ts
