+++ packages/common/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@web3modal/core",
-  "version": "3.2.1",
+  "name": "@web3modal/common",
+  "version": "3.2.0",
   "type": "module",
   "main": "./dist/esm/index.js",
   "types": "./dist/types/index.d.ts",
@@ -17,8 +17,7 @@
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx"
   },
   "dependencies": {
-    "@web3modal/common": "3.2.0",
-    "valtio": "1.11.2"
+    "dayjs": "1.11.10"
   },
   "keywords": [
     "web3",
