+++ packages/common/readme.md
@@ -1,0 +1,11 @@
+#### 📚 [Documentation](https://docs.walletconnect.com/2.0/web3modal/about)
+
+#### 🔗 [Website](https://web3modal.com)
+
+# Web3Modal
+
+Your on-ramp to web3 multichain. Web3Modal is a versatile library that makes it super easy to connect users with your Dapp and start interacting with the blockchain.
+
+<p align="center">
+  <img src="./.github/assets/header.png" alt="" border="0">
+</p>
