+++ packages/common/src/utils/TypeUtil.ts
@@ -1,0 +1,63 @@
+export type TransactionStatus = 'confirmed' | 'failed' | 'pending'
+
+export type TransactionDirection = 'in' | 'out' | 'self'
+
+export interface Transaction {
+  id: string
+  metadata: TransactionMetadata
+  transfers: TransactionTransfer[]
+}
+
+export interface TransactionMetadata {
+  operationType: string
+  hash: string
+  minedAt: string
+  sentFrom: string
+  sentTo: string
+  status: TransactionStatus
+  nonce: number
+}
+
+export interface TransactionTransfer {
+  fungible_info?: {
+    name?: string
+    symbol?: string
+    icon?: {
+      url: string
+    }
+  }
+  nft_info?: TransactionNftInfo
+  direction: TransactionDirection
+  quantity: TransactionQuantity
+  value?: number
+  price?: number
+}
+
+export interface TransactionNftInfo {
+  name?: string
+  content?: TransactionContent
+  flags: TransactionNftInfoFlags
+}
+
+export interface TransactionNftInfoFlags {
+  is_spam: boolean
+}
+
+export interface TransactionContent {
+  preview?: TransactionPreview
+  detail?: TransactionDetail
+}
+
+export interface TransactionPreview {
+  url: string
+  content_type?: null
+}
+
+export interface TransactionDetail {
+  url: string
+  content_type?: null
+}
+
+export interface TransactionQuantity {
+  numeric: string
+}
