+++ packages/common/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "composite": true,
+    "outDir": "./dist/esm",
+    "declarationDir": "./dist/types"
+  },
+  "extends": "../../tsconfig.json",
+  "include": ["index.ts", "src", "tests"]
+}
