+++ packages/ui/index.ts
@@ -49,6 +49,7 @@ export * from './src/layout/wui-separator/index.js'
 
 export { initializeTheming, setColorTheme, setThemeVariables } from './src/utils/ThemeUtil.js'
 export { UiHelperUtil } from './src/utils/UiHelperUtil.js'
+export { TransactionUtil } from './src/utils/TransactionUtil.js'
 export { customElement } from './src/utils/WebComponentsUtil.js'
 
 export type { IconType, VisualType } from './src/utils/TypeUtil.js'
