+++ packages/ui/package.json
@@ -21,6 +21,7 @@
   },
   "devDependencies": {
     "@types/qrcode": "1.5.4",
+    "@web3modal/common": "3.2.0",
     "eslint-plugin-lit": "1.9.1",
     "eslint-plugin-wc": "2.0.4"
   },
